python/requirements.txt
attributedict
web3py
requests